Added a default header and footer
When creating an HTML file, before this commit, the system was only
adding a simple body tag before and after the text that was converted
to HTML. This was not adequate so I added two resources with a default
header and footer that would be added in if no header and footer file
was selected.

The next step will be to change the title tag to match the text that
is given in the title box. This code will be added to the main window
class file when it is created and shifted later once the code has been
refactored which should really happen in one of the commits sometime
soon before the code becomes unmaintainable.